pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "jobhound"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "anthropic>=0.25",
    "textual>=0.52",
    "fastmcp>=0.4",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio", "respx>=0.21"]

[project.scripts]
jobhound = "jobhound.daemon:main"
jobhound-tui = "tui.app:main"
jobhound-server = "jobhound.mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["jobhound*", "tui*"]
